assumptions: add tar file format
The tar file format is block aligned:

    A tar archive file contains a series of blocks. Each block contains
    BLOCKSIZE bytes. Although this format may be thought of as being on
    magnetic tape, other media are often used.
    # See https://www.gnu.org/software/tar/manual/html_node/Standard.html

I have not found a way to change the BLOCKSIZE to something different
than 512 bytes. So I assume that 512 bytes is the only valid blocksize
in the tar on-disk format.

Also note: That there are different formats of the tar file format.  The
python library already lists 3 different formats:

    * read/write support for the POSIX.1-1988 (ustar) format.
    * read/write support for the GNU tar format
    * read/write support for the POSIX.1-2001 (pax) format.
    See https://docs.python.org/3.6/library/tarfile.html

Signed-off-by: Stefan Lengfeld <contact@stefanchrist.eu>