use new nomenclature 'hash-alg' instead of 'checksum-type'
Use '--hash-alg' instead of '--checksum-type' for the commandline
option, because it's more precise. Standard nomenclature would be hash
algorithm and message digest (for the checksum). The tool 'casync' uses
option '--digest' for the hash algorithm type, but that's slightly wrong
because that's the name for the checksum value/message digest, not the
hash function.

Signed-off-by: Stefan Lengfeld <contact@stefanchrist.eu>